Updated the model, expanded the QuoteServer, started a modelUpdater
Updated the model. It should now be in its final state, or failing that very near final state.

Expanded the QuoteServer to retrive additional information from yahoo finance. It now collects the following info as well:
	- Company name
	- Closing price
	- Stock exchnage listed on.
I can add anything else we feel we need at a later date

Started work on a modelUpdater, that handles updating the model with the lastest info from yahoo finance. This is not complete but is functional in its current state